requirements.txt
kafka-python
requests
python-dotenv
duckdb
pandas
redis
fastapi
uvicorn
bs4
pymongo
uvicorn[standard]